Fix type parameters cannot be determined errors.
Casting to an object before casting to a primitive data type seems to
fix the compiler-time error:
type parameters of <T>T cannot be determined; no unique maximal instance
exists for type variable T with upper bounds int,java.lang.Object

+++ src/no/runsafe/toybox/command/SpawnHorse.java
@@ -43,7 +43,7 @@ public class SpawnHorse extends PlayerCommand
 			if (variant == null)
 				variant = this.getRandomHorseVariant();
 
-			int count = parameters.getValue(COUNT);
+			int count = (Integer) parameters.getValue(COUNT);
 
 			for (int i = 0; i < count; ++i)
 				this.horseSpawner.spawnHorse(executor.getLocation(), type, variant, (Boolean)  parameters.getValue(TAME));
